@@ -14,7 +14,9 @@
 /** @type {import('@docusaurus/plugin-content-docs').SidebarsConfig} */
 const sidebars = {
   // By default, Docusaurus generates a sidebar from the docs folder structure
-  tutorialSidebar: [{type: 'autogenerated', dirName: '.'}],
+  tutorialSidebar: [{ type: 'autogenerated', dirName: '.' }],
+  v1Sidebar: [{ type: 'autogenerated', dirName: 'v1' }],
+  v2Sidebar: [{ type: 'autogenerated', dirName: 'v2' }],
 
   // But you can create a sidebar manually
   /*
@@ -30,4 +32,4 @@ const sidebars = {
    */
 };
 
-module.exports = sidebars;
+export default sidebars;
